refactor(api): genres → src/api modules (Effort 1)
Mechanical move of the genres feature from src/hooks/queries/genres to
src/api/genres, adding queryOptions factories for the query hooks. No behavior change.
