+++ src/api/genres/types.ts
@@ -1,0 +1,8 @@
+export type Genre = {
+  id: string;
+  name: string;
+};
+
+export const genresKeys = {
+  all: () => ["genres"] as const,
+};
