+++ src/api/genres/useGenres.ts
@@ -1,13 +1,8 @@
-import { useQuery } from "@tanstack/react-query";
+import { queryOptions, useQuery } from "@tanstack/react-query";
 import { supabase } from "@/integrations/supabase/client";
+import { Genre, genresKeys } from "./types";
 
-// Query key factory
-export const genresKeys = {
-  all: ["genres"] as const,
-};
-
-// Business logic function
-async function fetchGenres(): Promise<Array<{ id: string; name: string }>> {
+export async function fetchGenres(): Promise<Genre[]> {
   const { data, error } = await supabase
     .from("music_genres")
     .select("id, name")
@@ -20,13 +15,16 @@ async function fetchGenres(): Promise<Array<{ id: string; name: string }>> {
   return data || [];
 }
 
-// Hook
-export function useGenresQuery() {
-  return useQuery({
-    queryKey: genresKeys.all,
+export function genresQuery() {
+  return queryOptions({
+    queryKey: genresKeys.all(),
     queryFn: fetchGenres,
-    staleTime: 10 * 60 * 1000, // 10 minutes - genres don't change often
+    staleTime: 10 * 60 * 1000,
   });
+}
+
+export function useGenresQuery() {
+  return useQuery(genresQuery());
 }
 
 export function useGenres() {
