+++ src/components/GenreBadge.test.tsx
@@ -1,9 +1,9 @@
 import { describe, expect, it, vi, beforeEach } from "vitest";
 import { render, screen } from "@testing-library/react";
 import { GenreBadge } from "./GenreBadge";
-import * as useGenresModule from "@/hooks/queries/genres/useGenres";
+import * as useGenresModule from "@/api/genres/useGenres";
 
-vi.mock("@/hooks/queries/genres/useGenres");
+vi.mock("@/api/genres/useGenres");
 
 describe("GenreBadge", () => {
   beforeEach(() => {
