+++ src/pages/admin/AddGenreDialog.tsx
@@ -1,5 +1,5 @@
 import { useForm } from "react-hook-form";
-import { useCreateGenreMutation } from "@/hooks/queries/genres/useCreateGenreMutation";
+import { useCreateGenreMutation } from "@/api/genres/useCreateGenreMutation";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
