+++ src/pages/admin/ArtistsManagement/AddArtistDialog.tsx
@@ -23,7 +23,7 @@ import {
 import { Music } from "lucide-react";
 import { useAuth } from "@/contexts/AuthContext";
 import { useUserPermissionsQuery } from "@/hooks/queries/auth/useUserPermissions";
-import { useGenresQuery } from "@/hooks/queries/genres/useGenres";
+import { useGenresQuery } from "@/api/genres/useGenres";
 import { useCreateArtistMutation } from "@/hooks/queries/artists/useCreateArtist";
 import { useUpdateArtistMutation } from "@/hooks/queries/artists/useUpdateArtist";
 import { GenreMultiSelect } from "./GenreMultiSelect";
